BancoDeDados_SQLite_Herana - Crude de CategoriaDao

+++ BancoDeDados_SQLite_Herana/app/src/main/java/com/example/bancodedados_sqlite_herana/Categoria.java
@@ -3,21 +3,21 @@ package com.example.bancodedados_sqlite_herana;
 import java.io.Serializable;
 
 public class Categoria implements Serializable {
-    private int id;
+    private Long id;
     private String nome;
 
     public Categoria() {
     }
-    public Categoria(int id, String nome) {
+    public Categoria(Long id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
